+++ hq/app/schedule/IamJob.scala
@@ -1,19 +1,23 @@
 package schedule
 
+import aws.AwsClients
+import com.amazonaws.services.identitymanagement.AmazonIdentityManagementAsync
 import com.amazonaws.services.sns.AmazonSNSAsync
 import com.gu.anghammarad.models.Notification
 import config.Config.getAnghammaradSNSTopicArn
 import model._
 import play.api.{Configuration, Logging}
-import schedule.IamNotifications.makeNotification
+import schedule.IamDisable.{deletePassword, disableAccessKey}
 import schedule.IamFlaggedUsers.getFlaggedCredentialsReports
+import schedule.IamNotifications.makeNotification
 import schedule.IamNotifier.send
+import schedule.IamUsersToDisable.usersToDisable
 import services.CacheService
 import utils.attempt.FailedAttempt
 
 import scala.concurrent.ExecutionContext
 
-class IamJob(enabled: Boolean, cacheService: CacheService, snsClient: AmazonSNSAsync, dynamo: Dynamo,config: Configuration)(implicit val executionContext: ExecutionContext) extends JobRunner with Logging {
+class IamJob(enabled: Boolean, cacheService: CacheService, snsClient: AmazonSNSAsync, dynamo: Dynamo,config: Configuration, iamClients: AwsClients[AmazonIdentityManagementAsync])(implicit val executionContext: ExecutionContext) extends JobRunner with Logging {
   override val id = "credentials report job"
   override val description = "Automated emails for old permanent credentials"
   override val cronSchedule: CronSchedule = CronSchedules.everyWeekDay
@@ -27,8 +31,10 @@ class IamJob(enabled: Boolean, cacheService: CacheService, snsClient: AmazonSNSA
     }
 
     def getCredsReport(cacheService: CacheService): Map[AwsAccount, Either[FailedAttempt, CredentialReportDisplay]] = cacheService.getAllCredentials
+
     val credsReport: Map[AwsAccount, Either[FailedAttempt, CredentialReportDisplay]] = getCredsReport(cacheService)
     logger.info(s"successfully collected credentials report for $id. Report is empty: ${credsReport.isEmpty}.")
+    val flaggedCredentials = getFlaggedCredentialsReports(credsReport, dynamo)
 
     def sendNotificationAndRecord(notification: Notification, users: Seq[IamAuditUser]): Unit = {
       for {
@@ -37,9 +43,18 @@ class IamJob(enabled: Boolean, cacheService: CacheService, snsClient: AmazonSNSA
       } yield ()
     }
 
-    makeNotification(getFlaggedCredentialsReports(credsReport, dynamo)).foreach { notification =>
+    // send warning and final notifications
+    makeNotification(flaggedCredentials).foreach { notification =>
       notification.warningN.foreach(sendNotificationAndRecord(_, notification.alertedUsers))
       notification.finalN.foreach(sendNotificationAndRecord(_, notification.alertedUsers))
+    }
+
+    // disable user if still vulnerable after notifications have been sent and send a final notification stating this
+    usersToDisable(flaggedCredentials).foreach { user =>
+      disableAccessKey(iamClients, user)
+      deletePassword(iamClients, user)
+      val alertAwsAccountThatUserDisabled: Notification = ???
+      send(alertAwsAccountThatUserDisabled, topicArn, snsClient, testMode)
     }
   }
 }
